Extract command into its own Command class
